Remove durationUs from MediaFormat.
Duration was originally included in MediaFormat to match the
framework class, but it actually doesn't make much sense. In
many containers there's no such thing as per-stream duration,
and in any case we don't really care. Setting the duration on
each format required excessive piping.

This change moves duration into SeekMap instead, which seems
to make a lot more sense because it's at the container level,
and because being able to seek is generally couplied with
knowing how long the stream is.

This change is also a step toward merging Format and MediaFormat
into a single class (because Format doesn't have a duration),
which is coming soon.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=114428500

+++ library/src/main/java/com/google/android/exoplayer/extractor/ogg/VorbisUtil.java
@@ -75,7 +75,7 @@ import java.util.Arrays;
 
     long version = headerData.readLittleEndianUnsignedInt();
     int channels = headerData.readUnsignedByte();
-    long sampleRate = headerData.readLittleEndianUnsignedInt();
+    int sampleRate = (int) headerData.readLittleEndianUnsignedInt();
     int bitrateMax = headerData.readLittleEndianInt();
     int bitrateNominal = headerData.readLittleEndianInt();
     int bitrateMin = headerData.readLittleEndianInt();
@@ -433,7 +433,7 @@ import java.util.Arrays;
 
     public final long version;
     public final int channels;
-    public final long sampleRate;
+    public final int sampleRate;
     public final int bitrateMax;
     public final int bitrateNominal;
     public final int bitrateMin;
@@ -442,7 +442,7 @@ import java.util.Arrays;
     public final boolean framingFlag;
     public final byte[] data;
 
-    public VorbisIdHeader(long version, int channels, long sampleRate, int bitrateMax,
+    public VorbisIdHeader(long version, int channels, int sampleRate, int bitrateMax,
         int bitrateNominal, int bitrateMin, int blockSize0, int blockSize1, boolean framingFlag,
         byte[] data) {
       this.version = version;
